initial commit
the sum of the initial project steps since midth of may 2024, consisting of 3 separate days of working on the code and testing the stuff, with only a few hours invested in total, guess this will change once i focus my efforts mostly on this project.

read readme for further details on program handling and current features.