devolepment
improved prototype element textbox,

began improving element resizing

unsure atm if resize adorners and textboxes are a good fit, because with tabs and enters, resizing is also possible, but clumsy, with adorners, it is flexible, but would need some more logic like scrollbars, or some way to resize on enters and tabs, which is blocked right now.

before finishing this, the next goal is implementing a rudimentary xml based data serialization feature, to
remove time costs when testing.